GCP Projects/day_05/ex_03/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
google-auth==2.0.0
requests==2.26.0